requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
matplotlib==3.8.2
google-generativeai==0.3.2
python-dotenv==1.0.0
gunicorn==21.2.0
